apps/wow-producer/requirements.txt
confluent-kafka
watchdog
python-dotenv
lz4